+++ libs/num/compare.go
@@ -48,3 +48,31 @@ func AbsV[T Signed](a T) T {
 	}
 	return a
 }
+
+// MaxAbs - get max value based on absolute values of abolute vals.
+func MaxAbs[T Signed](vals ...T) T {
+	var r, m T
+	for _, v := range vals {
+		av := v
+		if av < 0 {
+			av *= -1
+		}
+		if av > m {
+			r = v
+			m = av // current max abs is av
+		}
+	}
+	return r
+}
+
+// CmpV compares 2 numberic values of any type, we attempt to cast T1 to T2 and back to see if that conversion
+// loses any information, if  no data is lost this way, we compare both values as T2, otherwise we compare both as T1.
+func CmpV[T1 Num, T2 Num](a T1, b T2) bool {
+	if a2 := T2(a); a == T1(a2) {
+		// cast to T2 can be cast back to T1 -> no information is lost, and so T2(a) can be safely compared to b
+		return a2 == b
+	}
+	// information was lost in T2(a), either a is float and b is int, or b is float32 vs a float64
+	// either way T1(b) is safe
+	return a == T1(b)
+}